fix setting active index on web on scroll
Corrected the logic for updating the active index during scroll events on the web client. Previously, the index was not syncing correctly with the scroll position due to onMomentumend scroll event was not firing properly onweb. so i use useDeived hook to set the active index correctly whenever scrollX shared value changes